requirements.txt
dnspython
streamlit